requirements.txt
os
streamlit
pandas
scikit-learn
pillow
joblib
